+++ packages/fuel-gauge/src/contract.test.ts
@@ -1,11 +1,7 @@
-import { generateTestWallet, seedTestWallet } from '@fuel-ts/account/test-utils';
 import { ErrorCode, FuelError } from '@fuel-ts/errors';
-import { expectToThrowFuelError } from '@fuel-ts/errors/test-utils';
-import { ASSET_A, ASSET_B } from '@fuel-ts/utils/test-utils';
 import type {
   TransactionRequestLike,
   TransactionResponse,
-  TransactionType,
   JsonAbi,
   ScriptTransactionRequest,
   TransferParams,
@@ -20,14 +16,21 @@ import {
   Provider,
   Contract,
   transactionRequestify,
-  FunctionInvocationResult,
   Wallet,
   ContractFactory,
   ZeroBytes32,
   FUEL_NETWORK_URL,
   Predicate,
   PolicyType,
+  buildFunctionResult,
 } from 'fuels';
+import {
+  generateTestWallet,
+  seedTestWallet,
+  expectToThrowFuelError,
+  ASSET_A,
+  ASSET_B,
+} from 'fuels/test-utils';
 
 import { FuelGaugeProjectsEnum, getFuelGaugeForcProject } from '../test/fixtures';
 
@@ -660,7 +663,12 @@ describe('Contract', () => {
     const {
       value: [resultA, resultB],
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } = await FunctionInvocationResult.build<any>(invocationScopes, response, true, contract);
+    } = await buildFunctionResult<any>({
+      funcScope: invocationScopes,
+      transactionResponse: response,
+      isMultiCall: true,
+      program: contract,
+    });
 
     expect(resultA.toHex()).toEqual(bn(num).add(1).toHex());
     expect(resultB.a).toEqual(!struct.a);
@@ -767,12 +775,12 @@ describe('Contract', () => {
       value: [resultA, resultB],
       transactionResult,
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } = await FunctionInvocationResult.build<any, TransactionType.Script>(
-      invocationScopes,
-      response,
-      true,
-      contract
-    );
+    } = await buildFunctionResult<any>({
+      funcScope: invocationScopes,
+      transactionResponse: response,
+      isMultiCall: true,
+      program: contract,
+    });
 
     expect(transactionResult.transaction.witnesses.length).toEqual(1);
     expect(transactionResult.transaction.witnesses[0].data).toEqual(signedTransaction);