+++ packages/fuels/src/cli-utils.ts
@@ -1,0 +1,1 @@
+export * from '@fuel-ts/utils/cli-utils';