+++ packages/program/src/functions/base-invocation-scope.ts
@@ -14,10 +14,15 @@ import * as asm from '@fuels/vm-asm';
 import { clone } from 'ramda';
 
 import { getContractCallScript } from '../contract-call-script';
-import type { ContractCall, InvocationScopeLike, TxParams } from '../types';
+import { buildDryRunResult, buildFunctionResult } from '../response';
+import type {
+  ContractCall,
+  InvocationScopeLike,
+  TxParams,
+  FunctionResult,
+  DryRunResult,
+} from '../types';
 import { assert, getAbisFromAllCalls } from '../utils';
-
-import { InvocationCallResult, FunctionInvocationResult } from './invocation-results';
 
 /**
  * Creates a contract call object based on the provided invocation scope.
@@ -359,7 +364,7 @@ export class BaseInvocationScope<TReturn = any> {
    *
    * @returns The result of the function invocation.
    */
-  async call<T = TReturn>(): Promise<FunctionInvocationResult<T>> {
+  async call<T = TReturn>(): Promise<FunctionResult<T>> {
     assert(this.program.account, 'Wallet is required!');
 
     const transactionRequest = await this.fundWithRequiredCoins();
@@ -369,12 +374,12 @@ export class BaseInvocationScope<TReturn = any> {
       estimateTxDependencies: false,
     });
 
-    return FunctionInvocationResult.build<T>(
-      this.functionInvocationScopes,
-      response,
-      this.isMultiCall,
-      this.program as AbstractContract
-    );
+    return buildFunctionResult<T>({
+      funcScope: this.functionInvocationScopes,
+      isMultiCall: this.isMultiCall,
+      program: this.program,
+      transactionResponse: response,
+    });
   }
 
   /**
@@ -382,7 +387,7 @@ export class BaseInvocationScope<TReturn = any> {
    *
    * @returns The result of the invocation call.
    */
-  async simulate<T = TReturn>(): Promise<InvocationCallResult<T>> {
+  async simulate<T = TReturn>(): Promise<DryRunResult<T>> {
     assert(this.program.account, 'Wallet is required!');
 
     if (!('populateTransactionWitnessesSignature' in this.program.account)) {
@@ -393,11 +398,15 @@ export class BaseInvocationScope<TReturn = any> {
     }
     const transactionRequest = await this.fundWithRequiredCoins();
 
-    const result = await this.program.account.simulateTransaction(transactionRequest, {
+    const callResult = await this.program.account.simulateTransaction(transactionRequest, {
       estimateTxDependencies: false,
     });
 
-    return InvocationCallResult.build<T>(this.functionInvocationScopes, result, this.isMultiCall);
+    return buildDryRunResult<T>({
+      funcScopes: this.functionInvocationScopes,
+      callResult,
+      isMultiCall: this.isMultiCall,
+    });
   }
 
   /**
@@ -405,32 +414,32 @@ export class BaseInvocationScope<TReturn = any> {
    *
    * @returns The result of the invocation call.
    */
-  async dryRun<T = TReturn>(): Promise<InvocationCallResult<T>> {
+  async dryRun<T = TReturn>(): Promise<DryRunResult<T>> {
     const { receipts } = await this.getTransactionCost();
 
     const callResult: CallResult = {
       receipts,
     };
 
-    return InvocationCallResult.build<T>(
-      this.functionInvocationScopes,
+    return buildDryRunResult<T>({
+      funcScopes: this.functionInvocationScopes,
       callResult,
-      this.isMultiCall
-    );
+      isMultiCall: this.isMultiCall,
+    });
   }
 
-  async get<T = TReturn>(): Promise<InvocationCallResult<T>> {
+  async get<T = TReturn>(): Promise<DryRunResult<T>> {
     const { receipts } = await this.getTransactionCost();
 
     const callResult: CallResult = {
       receipts,
     };
 
-    return InvocationCallResult.build<T>(
-      this.functionInvocationScopes,
+    return buildDryRunResult<T>({
+      funcScopes: this.functionInvocationScopes,
       callResult,
-      this.isMultiCall
-    );
+      isMultiCall: this.isMultiCall,
+    });
   }
 
   getProvider(): Provider {